signingscript: run `osslsigncode verify` after authenticode signing
Check in root CAs from Windows 7 so we can verify that the signature and
timestamp cross-signature properly chain to trusted roots.

XXX The CA list appears incomplete and I've been unable to find / export
an authoritative and complete list of trusted CAs and their purposes so
maybe this is hopeless.